Ch. 1.3: Simple typo fix + Grammar suggestion in table (#67)
* Simple typo fix

I believe this is a typo. "The cycle ... require multiple iterations." becomes "The cycle ... requires multiple iterations."

* Ch. 1.3: Grammar suggestion in kable

Removed the extra "be" in "... should be at least be...". Also, I'm thinking since 'Date' isn't referred to as a column name here it would make more sense grammatically to refer to 'Dates', right?