Expand quant/qual defintions
I wanted to prove a little more context for the distinction between quantitative and qualitative data in this section, as well as return to the idea of ordered and unordered sets. I also changed the unordered example from red/green/blue to math/science/history as technically there is an ordering of the colors by wavelength.